requirements.txt
python-dotenv>=1.1.1
requests>=2.31.0